requirements.txt
# Template Builder Backend
# Python dependencies

flask>=2.0.0
flask-cors>=4.0.0
python-docx>=0.8.11
openai>=1.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
tiktoken>=0.5.0
tenacity>=8.2.0
python-dateutil>=2.8.2
pytest>=7.0.0
databricks-sql-connector>=3.0.0
pdfplumber>=0.10.0
databricks-sdk>=0.1.0

